% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_create_datebase_for_given_word.R
\name{create_database}
\alias{create_database}
\title{create_database}
\usage{
create_database(word, lemma_table = "new_table", dbname = "medline",
  host = "localhost", port = 5432, user = "", password = "")
}
\arguments{
\item{word}{Character. Word which is analyzing.It's nesesary to add
sufix after "_" with part of speech tag. List of all tag and thier expands
is available on site:
\url{http://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}.}

\item{lemma_table}{Character. Name of table which will be created.}

\item{dbname}{Character. Information about PostgreSQL database
connection. Database name.}

\item{host}{Character. Information about PostgreSQL database
connection. Host.}

\item{port}{Numeric. Information about PostgreSQL database connection.
Port number.}

\item{user}{Character. Information about PostgreSQL database
connection. PostgreSQL username.}

\item{password}{Character. Information about PostgreSQL database
connection. PostgreSQL user's password.}
}
\value{
There is no output. \cr
This function create table in postgres with abstracts which indlude \code{word}.
}
\description{
Create a new table with abstracts which indlude search word. To correct work
this function needs 2 tables in PostgreSQL : \cr
library - word_id:INTEGER, word:TEXT \cr
abstracts - pmid:INTEGER, words:INTEGER[] \cr
}
\examples{
#create_database("cancer_NN")
}
