% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_datebase_to_table.R
\name{create_table_with_count_words}
\alias{create_table_with_count_words}
\title{Create table with count all words in analyzing abstracts.}
\usage{
create_table_with_count_words(lemma_table = "new_table", json,
  first_analyzed_date = "2000-01-01", last_analyzed_date = "2016-12-01")
}
\arguments{
\item{lemma_table}{Character. Name of source table or table from \code{\link{create_database}}.}

\item{json}{"json varible" or json file. JSON need to indlude: \cr
\code{dbname}      Character. Database name. \cr
\code{host}        Character. Host. \cr
\code{port}        Numeric. Port number. \cr \cr
And could indlude: \cr \cr
\code{user}        Character. PostgreSQL username. \cr
\code{password}    Character. PostgreSQL user's password. \cr}

\item{first_analyzed_date}{Date : YYYY-MM-DD. Month to start analysis}

\item{last_analyzed_date}{Date : YYYY-MM-DD. Month to stop analysis}
}
\value{
Table with count words from \code{lemma_table}.
}
\description{
Create table with count all words in analyzing abstracts.
}
\examples{
\dontrun{
create_table_with_count_words("new_table")
}
}
