% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_datebase_to_table.R
\name{create_data_to_visualization}
\alias{create_data_to_visualization}
\title{Opis funkcji}
\usage{
create_data_to_visualization(word, lemma_table, table_exist = F,
  output_file = "output_file", first_analyzed_date = "2000-01-01",
  last_analyzed_date = "2016-12-01", dictionary_exist = T,
  dbname = "medline", host = "localhost", port = 5432, user = "",
  password = "", save_to_file = T)
}
\arguments{
\item{word}{Character. Word which is analyzing.Its nesesary to add
sufix after "_" with part of speech tag. List of all tag and their expands
is available on site:
\url{http://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}.}

\item{lemma_table}{Character. Name of table which will be created or
existing table from \code{\link{create_database}}}

\item{table_exist}{Logical. If FALSE, function automatically  create new
table with name: \code{lemma_table}}

\item{output_file}{Character. Name of output file.}

\item{first_analyzed_date}{Date : YYYY-MM-DD. Month to start analysis}

\item{last_analyzed_date}{Date : YYYY-MM-DD. Month to stop analysis}

\item{dictionary_exist}{Logical. If you have a dictionary in file it
should TRUE. Otherwise function prepare this file, but this take a lot of
time.}

\item{dbname}{Character. Information about PostgreSQL database
connection. Database name.}

\item{host}{Character. Information about PostgreSQL database
connection. Host.}

\item{port}{Integer. Information about PostgreSQL database connection.
Port number.}

\item{user}{Character. Information about PostgreSQL database
connection. PostgreSQL username.}

\item{password}{Character. Information about PostgreSQL database
connection. PostgreSQL user's password.}

\item{save_to_file}{Logical. If TRUE create a output file with name:
\code{output_file}}
}
\value{
Table with count words in every time interval (default - year) in
  abstracts with \code{word}.
}
\description{
Opis funkcji
}
\examples{
#create_data_to_visualization()
}
