% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_create_datebase_for_given_word.R
\name{create_database}
\alias{create_database}
\title{create_database}
\usage{
create_database(word, lemma_table = "new_table", json)
}
\arguments{
\item{word}{Character. Word which is analyzing.Its nesesary to add
sufix after "_" with part of speech tag. List of all tag and their expands
is available on site:
\url{http://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}.}

\item{lemma_table}{Character. Name of table which will be created.}

\item{json}{"json varible" or json file. JSON need to indlude: \cr
dbname      Character. Database name. \cr
host        Character. Host. \cr
port        Numeric. Port number. \cr \cr
And could indlude: \cr \cr
user        Character. PostgreSQL username. \cr
password    Character. PostgreSQL user's password. \cr}
}
\value{
There is no output. \cr
This function create table in postgres with abstracts which include \code{word}.
}
\description{
Create a new table with abstracts which include search word. To correct work
this function needs 2 tables in PostgreSQL : \cr
library - word_id:INTEGER, word:TEXT \cr
abstracts - pmid:INTEGER, words:INTEGER[] \cr
}
\examples{
\dontrun{
create_database("cancer_NN")
}
}
