% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_datebase_to_table.R
\name{create_table_with_count_words}
\alias{create_table_with_count_words}
\title{Create table with count all words in analising abstracts.}
\usage{
create_table_with_count_words(lemma_table = "new_table", dbname = "medline",
  host = "localhost", port = 5432, user = "username",
  password = "password", first_analyzed_date = "2000-01-01",
  last_analyzed_date = "2016-12-01")
}
\arguments{
\item{lemma_table}{Character. Name of source table or table from \code{\link{create_database}}.}

\item{dbname}{Character. Information about PostgreSQL database
connection. Database name.}

\item{host}{Character. Information about PostgreSQL database
connection. Host.}

\item{port}{Integer. Information about PostgreSQL database connection.
Port number.}

\item{user}{Character. Information about PostgreSQL database
connection. PostgreSQL username.}

\item{password}{Character. Information about PostgreSQL database
connection. PostgreSQL user's password.}

\item{first_analyzed_date}{Date : YYYY-MM-DD. Month to start analysis}

\item{last_analyzed_date}{Date : YYYY-MM-DD. Month to stop analysis}
}
\value{
Table with count words from \code{lemma_table}.
}
\description{
Create table with count all words in analising abstracts.
}
\examples{
#create_table_with_count_words("new_table")
}
