% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_datebase_to_table.R
\name{download_dictionary}
\alias{download_dictionary}
\title{Create table with dictionary of all words.}
\usage{
download_dictionary(dbname = "medline", host = "localhost", port = 5432,
  user = "username", password = "password")
}
\arguments{
\item{dbname}{Character. Information about PostgreSQL database
connection. Database name.}

\item{host}{Character. Information about PostgreSQL database
connection. Host.}

\item{port}{Integer. Information about PostgreSQL database connection.
Port number.}

\item{user}{Character. Information about PostgreSQL database
connection. PostgreSQL username.}

\item{password}{Character. Information about PostgreSQL database
connection. PostgreSQL user's password.}
}
\value{
Table with all words existing in source data base.
}
\description{
Create table with dictionary of all words.
}
\examples{
#download_dictionary()
}
