% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_datebase_to_table.R
\name{preper_data_to_ANNA}
\alias{preper_data_to_ANNA}
\title{Parse and clearing data to visualisation in shiny-app ANNA.}
\usage{
preper_data_to_ANNA(input = "input", output_file = "output_file",
  dictionary = "dictionary", save_to_file = T)
}
\arguments{
\item{input}{Table. Table with count words from \code{\link{create_table_with_count_words}}.}

\item{output_file}{character. Opis.}

\item{dictionary}{character. Location and name of dictionary file.}

\item{save_to_file}{Logical. If TRUE create a output file with name:
\code{output_file}}
}
\value{
Table with count words in every time interval (default - year)
}
\description{
Parse and clearing data to visualisation in shiny-app ANNA.
}
\examples{
\dontrun{
preper_data_to_ANNA()
}
}
